Added enhancement to /api/health-impact-function-groups to provide the appropriate incidence and prevalence dataset and year for each function IF the default incidence/prevalence dataset and population year are provided as query string parameters.  For example: /health-impact-function-groups/3?incidencePrevalenceDataset=39&popYear=2020
Also added incidence_dataset_name and prevalence_dataset_name to the HIF definition returned by this function. The incidence and prevalence dataset_name and year should be displayed in the Value of Effects screen. The incidence, incidenceYear, prevalence, and prevalenceYear fields should  be included in the task when submitted.

Included initial support for the hifTaskUuid parameter that should be included when submitting a valuation task.